requirements.txt
matplotlib
numpy
Pillow
scipy
tqdm
